HTTPRequestUtils deleted(transferred to the main class file)
