+++ Makefile
@@ -12,7 +12,6 @@ SRCS = src/main.cpp \
 	src/utils.cpp \
 	src/HTTPRequest.cpp \
 	src/HTTPResponse.cpp \
-	src/HTTPRequestUtils.cpp \
 	src/Router.cpp \
 	src/server_utils.cpp \
 	src/Server.cpp \
